Extract dispatch from client
